Fix harmless compiler warning.
FossilOrigin-Name: 88b39c46c15ab6bdd0ac9536ddb6fb09862d578d0f691c57d039b0a906f1c25c

+++ ext/rtree/geopoly.c
@@ -286,7 +286,7 @@ static GeoPoly *geopolyFuncParam(
     int nVertex;
     nVertex = (a[1]<<16) + (a[2]<<8) + a[3];
     if( (a[0]==0 || a[0]==1)
-     && (nVertex*2*sizeof(GeoCoord) + 4)==nByte
+     && (nVertex*2*sizeof(GeoCoord) + 4)==(unsigned int)nByte
     ){
       p = sqlite3_malloc64( sizeof(*p) + (nVertex-1)*2*sizeof(GeoCoord) );
       if( p==0 ){
